backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
python-multipart
requests
python-dotenv
google-generativeai
gspread
oauth2client
python-telegram-bot
Pillow
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
